refactors utils module
